sandbox: Split the API metrics interceptor from the naming.
