+++ ledger/sandbox/src/test/suite/scala/com/digitalasset/platform/apiserver/MetricsInterceptorSpec.scala
@@ -14,9 +14,6 @@ import com.digitalasset.grpc.adapter.server.akka.ServerAdapter
 import com.digitalasset.grpc.adapter.utils.implementations.AkkaImplementation
 import com.digitalasset.grpc.sampleservice.Responding
 import com.digitalasset.ledger.api.testing.utils.AkkaBeforeAndAfterAll
-import com.digitalasset.ledger.api.v1.active_contracts_service.ActiveContractsServiceGrpc
-import com.digitalasset.ledger.api.v1.command_service.CommandServiceGrpc
-import com.digitalasset.ledger.api.v1.command_submission_service.CommandSubmissionServiceGrpc
 import com.digitalasset.platform.apiserver.MetricsInterceptorSpec._
 import com.digitalasset.platform.hello.HelloServiceGrpc.HelloService
 import com.digitalasset.platform.hello.{HelloRequest, HelloResponse, HelloServiceGrpc}
@@ -33,55 +30,6 @@ import scala.concurrent.duration.{DurationInt, FiniteDuration}
 import scala.concurrent.{ExecutionContext, Future}
 
 final class MetricsInterceptorSpec extends AsyncFlatSpec with AkkaBeforeAndAfterAll with Matchers {
-
-  behavior of "MetricsInterceptor.camelCaseToSnakeCase"
-
-  import MetricsInterceptor.camelCaseToSnakeCase
-
-  it should "leave an empty string unchanged" in {
-    camelCaseToSnakeCase("") shouldBe ""
-  }
-
-  it should "leave a snake_cased string unchanged" in {
-    camelCaseToSnakeCase("snake_cased") shouldBe "snake_cased"
-  }
-
-  it should "remove the capitalization of the first letter" in {
-    camelCaseToSnakeCase("Camel") shouldBe "camel"
-  }
-
-  it should "turn a single capital letter into a an underscore followed by a lower case letter" in {
-    camelCaseToSnakeCase("CamelCase") shouldBe "camel_case"
-    camelCaseToSnakeCase("camelCase") shouldBe "camel_case"
-  }
-
-  it should "keep acronyms together and change their capitalization as a single unit" in {
-    camelCaseToSnakeCase("DAML") shouldBe "daml"
-    camelCaseToSnakeCase("DAMLFactory") shouldBe "daml_factory"
-    camelCaseToSnakeCase("AbstractDAML") shouldBe "abstract_daml"
-    camelCaseToSnakeCase("AbstractDAMLFactory") shouldBe "abstract_daml_factory"
-    camelCaseToSnakeCase("AbstractDAMLProxyJVMFactory") shouldBe "abstract_daml_proxy_jvm_factory"
-  }
-
-  it should "treat single letter words intelligently" in {
-    camelCaseToSnakeCase("ATeam") shouldBe "a_team"
-    camelCaseToSnakeCase("TeamA") shouldBe "team_a"
-    camelCaseToSnakeCase("BustAMove") shouldBe "bust_a_move"
-
-    // the following is mostly to document a reasonable short-coming:
-    // a single letter word followed by an acronym will be detected as a single acronym
-    camelCaseToSnakeCase("AJVMHeap") shouldBe "ajvm_heap"
-  }
-
-  behavior of "MetricsInterceptor.nameFor"
-
-  import MetricsInterceptor.nameFor
-
-  it should "produce the expected name for a selection of services" in {
-    nameFor(CommandServiceGrpc.METHOD_SUBMIT_AND_WAIT.getFullMethodName) shouldBe "daml.lapi.command_service.submit_and_wait"
-    nameFor(CommandSubmissionServiceGrpc.METHOD_SUBMIT.getFullMethodName) shouldBe "daml.lapi.command_submission_service.submit"
-    nameFor(ActiveContractsServiceGrpc.METHOD_GET_ACTIVE_CONTRACTS.getFullMethodName) shouldBe "daml.lapi.active_contracts_service.get_active_contracts"
-  }
 
   behavior of "MetricsInterceptor"
